% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_file.R
\name{add_file}
\alias{add_file}
\title{Adds the \code{file} column to a data frame}
\usage{
add_file(data, parent, ext = ".rds")
}
\arguments{
\item{data}{A data frame.}

\item{parent}{A directory.}

\item{ext}{An extension.}
}
\value{
A data frame.
}
\description{
Adds the \code{file} column to a data frame
}
\examples{
data <- tibble::tibble(chunk = 1)
add_file(data, parent = tempdir())
}
