% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dchunkr-package.R
\docType{package}
\name{dchunkr-package}
\alias{dchunkr}
\alias{dchunkr-package}
\title{dchunkr: Work with Chunks of Data}
\description{
Helps to work with chunks of data in parallel and to cache the results of each chunk. It's basic approach to handling somewhat large datasets and long runtimes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/2DegreesInvesting/dchunkr}
  \item \url{https://2degreesinvesting.github.io/dchunkr/}
  \item \url{https://github.com/maurolepore/dchunkr}
  \item \url{https://maurolepore.github.io/dchunkr/}
  \item Report bugs at \url{https://github.com/maurolepore/dchunkr/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID})

Other contributors:
\itemize{
  \item 2 Degrees Investing Initiative \email{contact@2degrees-investing.org} [copyright holder, funder]
}

}
\keyword{internal}
