% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_chunk.R
\name{nest_chunk}
\alias{nest_chunk}
\title{Nest a data frame by chunks containing all elements of a group}
\usage{
nest_chunk(data, .by, chunks)
}
\arguments{
\item{data}{A data frame.}

\item{.by}{A column which values should be.}

\item{chunks}{Integer. Number of chunks.}
}
\value{
A nested data frame.
}
\description{
Nest a data frame by chunks containing all elements of a group
}
\examples{
data <- tibble::tibble(id = c(1, 1, 1, 2, 3))
out <- nest_chunk(data, .by = "id", chunks = 3)
out$data
}
