% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_rows.R
\name{order_rows}
\alias{order_rows}
\title{Order the rows of a data frame}
\usage{
order_rows(data, .fun = c("identity", "rev", "sample"))
}
\arguments{
\item{data}{Data frame.}

\item{.fun}{Character. A function name.}
}
\value{
Data frame.
}
\description{
Order the rows of a data frame
}
\examples{
withr::local_seed(123)

data <- tibble::tibble(x = 1:5)
order_rows(data)
order_rows(data, "rev")
order_rows(data, "sample")
}
