% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_undone.R
\name{pick_undone}
\alias{pick_undone}
\title{Add and filter (un)done files in a data frame}
\usage{
pick_undone(data)
}
\arguments{
\item{data}{A data frame with the column \code{file}.}
}
\value{
A data frame.
}
\description{
Add and filter (un)done files in a data frame
}
\examples{
data <- tibble::tibble(file = c(
  withr::local_tempdir(pattern = "exists_"),
  withr::local_tempfile(pattern = "doesnt_exist_")
))
data

pick_undone(data)
}
